requirements.txt
# algorithms
numpy==1.21.2
scipy==1.7.1
pandas==1.3.3
nltk==3.6.3
snakecase==1.0.1
anytree==2.8.0
six==1.16.0
strsim==0.0.3
networkx==2.6.3
chardet==4.0.0
python-Levenshtein==0.12.2
PuLP==2.5.1
pyemd==0.5.1
# data loading
python-dateutil==2.8.2